apps/learner-profile-api/pyproject.toml
[project]
name = "learner-profile-api"
version = "0.1.0"
description = "Learner Profile API - profile CRUD, onboarding, PHM sync, progressive profiling"
readme = { text = "", content-type = "text/markdown" }
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "redis>=5.0.0",
    "sqlmodel>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.14.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "fakeredis>=2.26.0",
    "respx>=0.22.0",
    "cryptography>=44.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "aiosqlite>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/learner_profile_api"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
